requirements.txt
python-telegram-bot[job-queue,ext]~=22.0
Flask~=2.3.0
aiohttp~=3.9.0
python-dotenv~=1.0.0